pyproject.toml
[tool.poetry]
name = "pytaiga-mcp"
version = "0.1.0"
description = "Taiga integration bridge for MCP"
authors = [
    "Jukka Aho <ahojukka5@gmail.com>",
    "Talha Orak <talhaorak.git@gmail.com>"
]
readme = "README.md"
packages = [{include = "pytaiga_mcp"}]

[tool.poetry.scripts]
pytaiga-mcp = "pytaiga_mcp.server:main"

[tool.poetry.dependencies]
python = "^3.10"
mcp = {extras = ["cli"], version = ">=0.1.0"}
fastapi = ">=0.104.0"
uvicorn = {extras = ["standard"], version = ">=0.23.0"}
python-dotenv = ">=0.15.0"
httpx = ">=0.25.0"
pydantic = ">=2.4.0"
pydantic-settings = ">=2.0.0"
tenacity = ">=8.2.0"
pytaigaclient = {git = "https://github.com/talhaorak/pyTaigaClient.git"}

[tool.poetry.group.dev.dependencies]
pytest = ">=7.3.1"
pytest-asyncio = ">=0.21.0"
pytest-cov = ">=4.1.0"
pytest-mock = ">=3.12.0"
responses = ">=0.23.1"
black = ">=23.3.0"
isort = ">=5.12.0"
mypy = ">=1.3.0"
flake8 = ">=6.0.0"
ruff = ">=0.1.0"
mkdocs = "^1.6.1"
mkdocs-material = "^9.7.0"
mkdocstrings = "^0.30.1"
mkdocstrings-python = "^1.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 100

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "unit: Run unit tests",
    "integration: Run integration tests",
    "slow: Mark tests as slow-running",
    "auth: Authentication tests",
    "core: Core functionality tests",
    "projects: Project-related tests",
    "epics: Epic-related tests",
    "user_stories: User story-related tests",
    "tasks: Task-related tests",
    "issues: Issue-related tests",
    "sprints: Sprint-related tests",
]
addopts = "-v"

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
warn_unreachable = true
strict_equality = true
pretty = true
show_error_context = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "pytaigaclient.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mcp.*"
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise from within except without exception chaining
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
